Create pymagicc.io module
