+++ pymagicc/__init__.py
@@ -8,8 +8,8 @@
 # License (https://creativecommons.org/licenses/by-nc-sa/3.0/)
 from ._version import get_versions
 from .core import MAGICC6, MAGICC7  # noqa
-from .io import MAGICCData, read_scen_file  # noqa
-from .scenarios import rcp26, rcp45, rcp60, rcp85, rcps, zero_emissions  # noqa
+from .io import MAGICCData   # noqa
+from .scenarios import rcp26, rcp45, rcp60, rcp85, rcps, zero_emissions, read_scen_file  # noqa
 
 __version__ = get_versions()["version"]
 del get_versions
